License fix: move async_utils to common
Technically, only `sync_to_async()` is used by datasets, but
probably it is easier to manage if all async utils stay together.

They only depend on the Python standard library, so no license issue.
