+++ src/libertem/io/dataset/__init__.py
@@ -3,7 +3,7 @@ import importlib
 from typing_extensions import Literal
 
 from libertem.io.dataset.base import DataSetException, DataSet
-from libertem.utils.async_utils import sync_to_async
+from libertem.common.async_utils import sync_to_async
 from libertem.executor.scheduler import Scheduler
